Factored out S3 stuff from CIBuild stuff
